requirements.txt
yt-dlp
Flask~=3.1.0
flask_babel~=4.0.0
Flask-Limiter~=3.12
python-dotenv~=1.1.0
gunicorn~=23.0.0
psutil~=7.0.0
ipaddress~=1.0.23
Werkzeug~=3.1.3
requests~=2.32.3
